shell: extract common init / shutdown
