+++ src/libtrx/game/shell/platform.h
@@ -1,0 +1,6 @@
+#pragma once
+
+// Isolated platform-sensitive initialization code
+
+void Shell_SetupHiDPI(void);
+void Shell_SetupLibAV(void);
