+++ src/libtrx/meson.build
@@ -249,7 +249,9 @@ sources = [
   'game/scaler.c',
   'game/shell/common.c',
   'game/shell/events.c',
+  'game/shell/flow.c',
   'game/shell/main.c',
+  'game/shell/platform.c',
   'game/sound/common.c',
   'game/stats.c',
   'game/ui/common.c',
