+++ src/tr1/game/sound.h
@@ -6,8 +6,6 @@
 
 #include <stdint.h>
 
-bool Sound_Init(void);
-void Sound_Shutdown(void);
 void Sound_UpdateEffects(void);
 void Sound_ResetEffects(void);
 void Sound_StopAmbientSounds(void);
