+++ src/tr2/game/game_string.h
@@ -1,6 +1,3 @@
 #pragma once
 
 #include <libtrx/game/game_string.h>
-
-void GameString_Init(void);
-void GameString_Shutdown(void);
