+++ src/tr2/game/sound.h
@@ -6,8 +6,5 @@
 
 #define SOUND_DEFAULT_PITCH 0x10000
 
-void Sound_Init(void);
-void Sound_Shutdown(void);
-
 void Sound_UpdateEffects(void);
 void Sound_EndScene(void);
